+++ test/main/presenter/llmProviderPresenter/acpFsHandler.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach } from 'vitest'
 import * as fs from 'fs/promises'
 import * as path from 'path'
 import * as os from 'os'
-import { AcpFsHandler } from '@/presenter/llmProviderPresenter/agent/acpFsHandler'
+import { AcpFsHandler } from '@/presenter/agentPresenter/acp'
 
 describe('AcpFsHandler', () => {
   let testDir: string
